- Added a check for an invalid index in AlterBookData

+++ Bookling/Bookling.Controller/LibraryDatabaseManager.cs
@@ -220,6 +220,9 @@ namespace Bookling.Controller
 
 		public void RemoveBook (int bookID) 
 		{
+			if (bookID < 0) {
+				throw new BookNotFoundException ();
+			}
 			try {
 				using (SqliteCommand command = new SqliteCommand (Connection)) {
 					command.CommandText = "DELETE FROM Books WHERE BookID = :id;";
